--- src/app/components/CardButton/index.tsx
@@ -1,3 +1,0 @@
-import CardButton from "./CardButton";
-
-export default CardButton;
