--- src/app/components/blockchain/ActiveNetwork/index.tsx
@@ -1,3 +1,0 @@
-import ActiveNetwork from "./ActiveNetwork";
-
-export default ActiveNetwork;
