--- src/app/components/blockchain/Disconnected/index.tsx
@@ -1,3 +1,0 @@
-import NotConnected from "./Disconnected";
-
-export default NotConnected;
