+++ src/app/page.tsx
@@ -2,7 +2,7 @@
 import { useAccount } from "wagmi";
 import { Text } from "@0xsequence/design-system";
 import Connected from "./components/blockchain/Connected";
-import NotConnected from "./components/blockchain/Disconnected";
+import NotConnected from "./components/blockchain/NotConnected";
 
 const HomePage = () => {
   const { isConnected, chain } = useAccount();
